LayoutTests/platform/ios-sim-deprecated/fast/events/spatial-navigation/snav-unit-overflow-and-scroll-in-direction-expected.txt

Start
End
PASS document.activeElement.getAttribute("id") is "start"
FAIL false should be true. Was false.
FAIL document.activeElement.getAttribute("id") should be end. Was start.
